moved POSSIBLE_USERS to rolloutUtils.ts
