+++ torchci/lib/drciUtils.ts
@@ -10,9 +10,6 @@ export const CPP_DOCS_URL = "/cppdocs/index.html";
 export const DRCI_COMMENT_END = `\n
 This comment was automatically generated by Dr. CI and updates every 15 minutes.
 <!-- drci-comment-end -->`;
-export const POSSIBLE_USERS = ["swang392", "janeyx99", "atalman", "clee2000", "seemethere",
-    "huydhn", "zengk95", "linbinyu", "suo", "dagitses", "malfet", "mehtanirav", "kit1980",
-    "b0noI", "weiwangmeta", "ZainRizvi", "DanilBaibak", "jeanschmidt", "osalpekar"];
 export const HUD_URL = "https://hud.pytorch.org/pr/";
