+++ torchci/lib/rolloutUtils.ts
@@ -1,0 +1,3 @@
+export const POSSIBLE_USERS = ["swang392", "janeyx99", "atalman", "clee2000", "seemethere",
+    "huydhn", "zengk95", "linbinyu", "suo", "dagitses", "malfet", "mehtanirav", "kit1980",
+    "b0noI", "weiwangmeta", "ZainRizvi", "DanilBaibak", "jeanschmidt", "osalpekar"];
